Make --rebuild work correctly again. (#111)
Add missing submodules to rebuild, consistently check for the existence
of build files, always attempt to build from the correct directory, and
try to build in the correct order. The glog project is not re-built as
it is very unlikely that changes will be made to it and it isn't
supported with all fuzzers (memory).