Use CC and CXX, check for clang version.
Currently cc and cxx are hard-wired to 'clang' and 'clang++', however
the user may wish to use another build of clang, like a local build or
something like 'clang-12'. Use CC and CXX when they have 'clang' in
them.

Test the clang version. In order to build llvm-13 libcxx clang version
10 or greater is required. Exit configuration if the detected version is
less than 10.

Output the configured settings for cc and cxx and their versions to
clarify the selected settings.

Fixes: #120