Build bzip2 correctly. (#124)
The bzip2 build is just a simple Makefile which overrides CC, CFLAGS,
and LDFLAGS. Patch up the Makefile not to do that. Note that the
oss-fuzz bzip2 build does not use the Makefile at all, instead manually
building and linking in the build.sh.

This change allows the i386 oss-fuzz configuration to work.