Avoid HAVE_ARCHIVE modifying interfaces
There is no need to guard the declaration of `set_data_is_tar_archive`
behind `HAVE_ARCHIVE`. It is known that it may or may not have an
effect. In addition, changing the public interface layout like this is
leading to issues.

The only use of `HAVE_ARCHIVE` should be to guard the code that uses the
`tarreader` in the `FaceLoader::set_raw_bytes` implementation. In
addition, the `FaceLoadIterator` constructor and destructor are moved
out of line to ensure they use the same declarations when creating and
destroying the `FaceLoader`.

Bug: https://issues.oss-fuzz.com/issues/444537258